code refactoring done
