--- domain-consumer/src/main/java/com/smallintro/springboot/domainconsumer/DomainConsumerApplication.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domainconsumer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainConsumerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DomainConsumerApplication.class, args);
-	}
-
-}
