+++ domain-consumer/src/main/java/io/github/smallintro/kafka/domainconsumer/DomainConsumerApplication.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domainconsumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainConsumerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DomainConsumerApplication.class, args);
+	}
+
+}
