--- domain-consumer/src/test/java/com/smallintro/springboot/domainconsumer/DomainConsumerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domainconsumer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DomainConsumerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
