+++ domain-consumer/src/test/java/io/github/smallintro/kafka/domainconsumer/DomainConsumerApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domainconsumer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainConsumerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
