--- domain-crawler/src/main/java/com/smallintro/springboot/domaincrawler/DomainCrawlerApplication.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domaincrawler;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainCrawlerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DomainCrawlerApplication.class, args);
-	}
-
-}
