--- domain-crawler/src/main/java/com/smallintro/springboot/domaincrawler/constant/KafkaConstants.java
@@ -1,5 +1,0 @@
-package com.smallintro.springboot.domaincrawler.constant;
-
-public interface KafkaConstants {
-    String ACTIVE_WEB_TOPIC = "web-domains-topic";
-}
