+++ domain-crawler/src/main/java/io/github/smallintro/kafka/domaincrawler/DomainCrawlerApplication.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domaincrawler;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainCrawlerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DomainCrawlerApplication.class, args);
+	}
+
+}
