+++ domain-crawler/src/main/java/io/github/smallintro/kafka/domaincrawler/constant/KafkaConstants.java
@@ -1,0 +1,5 @@
+package io.github.smallintro.kafka.domaincrawler.constant;
+
+public interface KafkaConstants {
+    String ACTIVE_WEB_TOPIC = "web-domains-topic";
+}
