--- domain-crawler/src/test/java/com/smallintro/springboot/domaincrawler/DomainCrawlerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domaincrawler;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DomainCrawlerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
