+++ domain-crawler/src/test/java/io/github/smallintro/kafka/domaincrawler/DomainCrawlerApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domaincrawler;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainCrawlerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
