--- domain-processor/src/main/java/com/smallintro/springboot/domainprocessor/DomainProcessorApplication.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domainprocessor;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainProcessorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DomainProcessorApplication.class, args);
-	}
-
-}
