+++ domain-processor/src/main/java/io/github/smallintro/kafka/domainprocessor/DomainProcessorApplication.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domainprocessor;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DomainProcessorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DomainProcessorApplication.class, args);
+	}
+
+}
