--- domain-processor/src/test/java/com/smallintro/springboot/domainprocessor/DomainProcessorApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.domainprocessor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DomainProcessorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
