+++ domain-processor/src/test/java/io/github/smallintro/kafka/domainprocessor/DomainProcessorApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.domainprocessor;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainProcessorApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
