--- order-processor/src/main/java/com/smallintro/springboot/orderprocessor/OrderProcessorApplication.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.orderprocessor;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderProcessorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderProcessorApplication.class, args);
-	}
-
-}
