--- order-processor/src/main/java/com/smallintro/springboot/orderprocessor/constant/KafkaConstants.java
@@ -1,11 +1,0 @@
-package com.smallintro.springboot.orderprocessor.constant;
-
-public interface KafkaConstants {
-
-    String KAFKA_NODE_ADDRESS = "127.0.0.1:9092";
-    String TOPIC = "kafka-string-msg-topic";
-    String GROUP_ID = "kafka-string-msg-group";
-
-    String TOPIC_JSON = "kafka-json-msg-topic";
-    String GROUP_ID_JSON = "kafka-json-msg-group";
-}
