+++ order-processor/src/main/java/io/github/smallintro/kafka/orderprocessor/OrderProcessorApplication.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.orderprocessor;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrderProcessorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OrderProcessorApplication.class, args);
+	}
+
+}
