+++ order-processor/src/main/java/io/github/smallintro/kafka/orderprocessor/constant/KafkaConstants.java
@@ -1,0 +1,11 @@
+package io.github.smallintro.kafka.orderprocessor.constant;
+
+public interface KafkaConstants {
+
+    String KAFKA_NODE_ADDRESS = "127.0.0.1:9092";
+    String TOPIC = "kafka-string-msg-topic";
+    String GROUP_ID = "kafka-string-msg-group";
+
+    String TOPIC_JSON = "kafka-json-msg-topic";
+    String GROUP_ID_JSON = "kafka-json-msg-group";
+}
