--- order-processor/src/test/java/com/smallintro/springboot/orderprocessor/OrderProcesserApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smallintro.springboot.orderprocessor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderProcesserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
