+++ order-processor/src/test/java/io/github/smallintro/kafka/orderprocessor/OrderProcesserApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.smallintro.kafka.orderprocessor;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderProcesserApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
